pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "comai"
dynamic = ["version"]
authors = [{ name = "Pedro Rico", email = "ricopinazo@gmail.com" }]
description = "AI powered console assistant"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Operating System :: MacOS",
    "Operating System :: Unix",
]
dependencies = [
    "typer-slim==0.12.3",
    "rich==13.7.1",
    "prompt-toolkit==3.0.43",
    "simple-term-menu==1.6.4",
    "langchain==0.1.17",
    "langchain-openai==0.1.6",
    "ollama==0.1.9",
]

[project.urls]
homepage = "https://github.com/ricopinazo/comai"
repository = "https://github.com/ricopinazo/comai"
issues = "https://github.com/ricopinazo/comai/issues"

[project.scripts]
comai = "comai.cli:app"

[project.optional-dependencies]
dev = [
    "pytest",
    "hatchling",
    "types-requests==2.31.0.20240406",
    "mypy==1.9.0",
    "black==24.4.0",
]

[tool.hatch.version]
path = "src/comai/__init__.py"

[tool.mypy]
ignore_missing_imports = true
